CerebrasApp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.4
pandas==2.1.3
openpyxl==3.1.2
supabase==2.4.5
stripe==8.10.0
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.5.0